requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-multipart
